fix: Handle unavailable commitment_update in case of a disconnected peer
In case the peer already received the accept message and looses the connection to the counterparty while processing the accept message the `commitment_update` will not be set.

The Option needs to be properly handled and an error is returned if no `commitment_update` is available. The accept message will be reprocessed upon reconnect to the counterparty.