fix: Do not panic if commitment update is missing
In the context of the LN-DLC protocol, if the peer has already
_received_ the accept message and loses the connection to the
counterparty while _processing_ the accept message the
`commitment_update` will not be set.

Therefore, the `Option` needs to be properly handled and an error
returned if no `commitment_update` is available.

The accept message will be reprocessed upon reconnection to the
counterparty.

Co-authored-by: Richard Holzeis <richard@holzeis.me>